examples/pyproject.toml
[tool.poetry]
name = "reya-examples"
version = "0.1.0"
description = ""
authors = ["Reya Labs"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typing-extensions = "^4.12.2"
eth-abi = "^5.1.0"
python-dotenv = "^1.0.1"
websocket-client = "^1.8.0"
web3 = "^6.20.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
